Move appdistro tests (#7318)
